fix: standardize dividend shares to Double type (BUGS3 #8)
DividendEvent.shares was a Long (supposedly in 1/10000 units) but the
repository stored it as Double directly and then multiplied by 10000
when reading, causing severe corruption. Now uses Double consistently
like the rest of the investment model.

+++ shared/src/commonMain/kotlin/com/financeapp/domain/model/Performance.kt
@@ -107,7 +107,7 @@ data class DividendEvent(
     val paymentDate: Long,
     val amount: Long, // Total dividend in cents
     val perShare: Long, // Dividend per share in cents
-    val shares: Long, // Number of shares at time of dividend (in 1/10000)
+    val shares: Double, // Number of shares at time of dividend
     val isReinvested: Boolean = false
 )
